Rename module: util.value.cell => util.cell
