+++ code/ui.datagrid/package.json
@@ -15,13 +15,13 @@
   },
   "dependencies": {
     "@babel/polyfill": "7.6.0",
-    "@platform/cache": "^0.1.0",
+    "@platform/cache": "0.1.0",
     "@platform/libs": "0.1.3",
     "@platform/react": "0.3.26",
     "@platform/ui.datagrid.types": "0.6.12",
     "@platform/ui.datagrid.util": "0.0.8",
     "@platform/ui.text": "0.3.0",
-    "@platform/util.value.cell": "0.10.3",
+    "@platform/util.cell": "0.0.0",
     "handsontable": "6.2.2"
   },
   "devDependencies": {
