+++ code/util.cell/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@platform/util.value.cell",
-  "version": "0.10.3",
+  "name": "@platform/util.cell",
+  "version": "0.0.0",
   "description": "Data types for working with grid cell cooridnates (eg 'A1').",
   "main": "lib/index",
   "types": "lib/index.d.ts",
