+++ code/util.cell/src/cell/cell.key.ts
