+++ code/util.cell/src/range/CellRange/CellRange.ts
