+++ code/util.cell/src/range/types.ts
