+++ code/util.cell/src/table/table.shift.TEST.ts
