+++ code/util.cell/src/table/table.shift.ts
