+++ code/util.cell/src/table/table.string.TEST.ts
