+++ code/util.cell/src/table/table.string.ts
