+++ code/util.cell/src/table/types.ts
